# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: I.D.A Immobilier API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties with filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, published, reserved, sold, rented, archived]
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: minArea
          schema:
            type: number
        - in: query
          name: maxArea
          schema:
            type: number
        - in: query
          name: rooms
          schema:
            type: integer
        - in: query
          name: bedrooms
          schema:
            type: integer
        - in: query
          name: agencyId
          schema:
            type: integer
        - in: query
          name: agentId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
...
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /properties/{id}/publish:
    patch:
      operationId: publishProperty
      tags: [properties]
      summary: Publish a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/media:
    get:
      operationId: listPropertyMedia
      tags: [media]
      summary: List media for a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropertyMedia"
    post:
      operationId: addPropertyMedia
      tags: [media]
      summary: Add media to a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyMediaInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyMedia"
  /properties/{id}/media/{mediaId}:
    delete:
      operationId: deletePropertyMedia
      tags: [media]
      summary: Delete a media item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /properties/{id}/assign:
    post:
      operationId: assignProperty
      tags: [properties]
      summary: Temporarily assign a property to another agent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyAssignmentInput"
      responses:
        "201":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyAssignment"
  /properties/{id}/similar:
    get:
      operationId: getSimilarProperties
      tags: [properties]
      summary: Get similar properties
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"

  # AGENCIES
...
components:
  schemas:
...
    Property:
      type: object
      required: [id, title, type, status, city, postalCode, address, agencyId, ownerAgentId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: ["string", "null"]
        type:
          type: string
          enum: [apartment, house, villa, land, commercial, garage, other]
        status:
          type: string
          enum: [draft, published, reserved, sold, rented, archived]
        # Location
        address:
          type: string
        postalCode:
          type: string
        city:
          type: string
        department:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        country:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        # Pricing
        salePrice:
          type: ["number", "null"]
        rentalPrice:
          type: ["number", "null"]
        charges:
          type: ["number", "null"]
        agencyFees:
          type: ["number", "null"]
        taxeFonciere:
          type: ["number", "null"]
        # Characteristics
        livingArea:
          type: ["number", "null"]
        landArea:
          type: ["number", "null"]
        carrezArea:
          type: ["number", "null"]
        rooms:
          type: ["integer", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        floor:
          type: ["integer", "null"]
        totalFloors:
          type: ["integer", "null"]
        yearBuilt:
          type: ["integer", "null"]
        orientation:
          type: ["string", "null"]
        heating:
          type: ["string", "null"]
        # Energy
        dpeRating:
          type: ["string", "null"]
        gesRating:
          type: ["string", "null"]
        annualEnergyCost:
          type: ["number", "null"]
        # Equipment (booleans)
        hasTerrace:
          type: boolean
        hasBalcony:
          type: boolean
        hasGarage:
          type: boolean
        hasParking:
          type: boolean
        hasGarden:
          type: boolean
        hasPool:
          type: boolean
        hasCellar:
          type: boolean
        hasElevator:
          type: boolean
        hasAirConditioning:
          type: boolean
        hasFiber:
          type: boolean
        # Descriptions
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        # SEO
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        # Relations
        agencyId:
          type: integer
        agencyName:
          type: ["string", "null"]
        ownerAgentId:
          type: integer
        currentAgentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        agentPhone:
          type: ["string", "null"]
        agentEmail:
          type: ["string", "null"]
        agentAvatarUrl:
          type: ["string", "null"]
        # Media
        mainImageUrl:
          type: ["string", "null"]
        mediaCount:
          type: integer
        viewCount:
          type: integer
        favoriteCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PropertyInput:
      type: object
      required: [title, type, address, postalCode, city]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [apartment, house, villa, land, commercial, garage, other]
        address:
          type: string
        postalCode:
          type: string
        city:
          type: string
        department:
          type: string
        region:
          type: string
        country:
          type: string
        salePrice:
          type: number
        rentalPrice:
          type: number
        charges:
          type: number
        agencyFees:
          type: number
        taxeFonciere:
          type: number
        livingArea:
          type: number
        landArea:
          type: number
        carrezArea:
          type: number
        rooms:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        floor:
          type: integer
        totalFloors:
          type: integer
        yearBuilt:
          type: integer
        orientation:
          type: string
        heating:
          type: string
        dpeRating:
          type: string
        gesRating:
          type: string
        annualEnergyCost:
          type: number
        hasTerrace:
          type: boolean
        hasBalcony:
          type: boolean
        hasGarage:
          type: boolean
        hasParking:
          type: boolean
        hasGarden:
          type: boolean
        hasPool:
          type: boolean
        hasCellar:
          type: boolean
        hasElevator:
          type: boolean
        hasAirConditioning:
          type: boolean
        hasFiber:
          type: boolean
        shortDescription:
          type: string
        fullDescription:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        agencyId:
          type: integer
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [draft, published, reserved, sold, rented, archived]
        type:
          type: string
        address:
          type: string
        postalCode:
          type: string
        city:
          type: string
        salePrice:
          type: number
        rentalPrice:
          type: number
        charges:
          type: number
        agencyFees:
          type: number
        taxeFonciere:
          type: number
        livingArea:
          type: number
        landArea:
          type: number
        carrezArea:
          type: number
        rooms:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        floor:
          type: integer
        totalFloors:
          type: integer
        yearBuilt:
          type: integer
        orientation:
          type: string
        heating:
          type: string
        dpeRating:
          type: string
        gesRating:
          type: string
        annualEnergyCost:
          type: number
        hasTerrace:
          type: boolean
        hasBalcony:
          type: boolean
        hasGarage:
          type: boolean
        hasParking:
          type: boolean
        hasGarden:
          type: boolean
        hasPool:
          type: boolean
        hasCellar:
          type: boolean
        hasElevator:
          type: boolean
        hasAirConditioning:
          type: boolean
        hasFiber:
          type: boolean
        shortDescription:
          type: string
        fullDescription:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
...
    PropertyMedia:
      type: object
      required: [id, propertyId, url, type, order, createdAt]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        url:
          type: string
        watermarkedUrl:
          type: ["string", "null"]
        type:
          type: string
          enum: [photo, video, floor_plan, pdf]
        caption:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
    PropertyMediaInput:
      type: object
      required: [url, type]
      properties:
        url:
          type: string
        type:
          type: string
          enum: [photo, video, floor_plan, pdf]
        caption:
          type: string
        order:
          type: integer

    # PROPERTY ASSIGNMENT
    PropertyAssignment:
      type: object
      required: [id, propertyId, ownerAgentId, temporaryAgentId, startDate, endDate]
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        ownerAgentId:
          type: integer
        temporaryAgentId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PropertyAssignmentInput:
      type: object
      required: [temporaryAgentId, startDate, endDate]
      properties:
        temporaryAgentId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string

    # LEAD SCHEMAS
...