# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * I.D.A Immobilier API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateStatus } from './propertyUpdateStatus';

export interface PropertyUpdate {
  title?: string;
  status?: PropertyUpdateStatus;
  type?: string;
  address?: string;
  postalCode?: string;
  city?: string;
  salePrice?: number;
  rentalPrice?: number;
  charges?: number;
  agencyFees?: number;
  taxeFonciere?: number;
  livingArea?: number;
  landArea?: number;
  carrezArea?: number;
  rooms?: number;
  bedrooms?: number;
  bathrooms?: number;
  floor?: number;
  totalFloors?: number;
  yearBuilt?: number;
  orientation?: string;
  heating?: string;
  dpeRating?: string;
  gesRating?: string;
  annualEnergyCost?: number;
  hasTerrace?: boolean;
  hasBalcony?: boolean;
  hasGarage?: boolean;
  hasParking?: boolean;
  hasGarden?: boolean;
  hasPool?: boolean;
  hasCellar?: boolean;
  hasElevator?: boolean;
  hasAirConditioning?: boolean;
  hasFiber?: boolean;
  shortDescription?: string;
  fullDescription?: string;
  metaTitle?: string;
  metaDescription?: string;
}